test: execute batch tests in both query protocol modes
